Move lib/utils/testhelpers.go to lib/utils/testutils
