+++ integration/port_forwarding_test.go
@@ -41,6 +41,7 @@ import (
 	"github.com/gravitational/teleport/lib/client"
 	"github.com/gravitational/teleport/lib/service/servicecfg"
 	"github.com/gravitational/teleport/lib/utils"
+	"github.com/gravitational/teleport/lib/utils/testutils"
 )
 
 func extractPort(svr *httptest.Server) (int, error) {
@@ -96,7 +97,7 @@ func testPingLocalServer(t *testing.T, port int, expectSuccess bool) {
 }
 
 func testPortForwarding(t *testing.T, suite *integrationTestSuite) {
-	invalidOSLogin := utils.GenerateLocalUsername(t)
+	invalidOSLogin := testutils.GenerateLocalUsername(t)
 
 	// Providing our own logins to Teleport so we can verify that a user
 	// that exists within Teleport but does not exist on the local node
