+++ lib/srv/regular/sshserver_unix_test.go
@@ -29,15 +29,15 @@ import (
 	"github.com/stretchr/testify/require"
 
 	tracessh "github.com/gravitational/teleport/api/observability/tracing/ssh"
-	"github.com/gravitational/teleport/lib/utils"
 	"github.com/gravitational/teleport/lib/utils/host"
+	"github.com/gravitational/teleport/lib/utils/testutils"
 )
 
 // BenchmarkRootExecCommand measures performance of running multiple exec requests
 // over a single ssh connection. The same test is run with and without host user
 // creation support to catch any performance degradation caused by user provisioning.
 func BenchmarkRootExecCommand(b *testing.B) {
-	utils.RequireRoot(b)
+	testutils.RequireRoot(b)
 
 	b.ReportAllocs()
 
@@ -66,7 +66,7 @@ func BenchmarkRootExecCommand(b *testing.B) {
 			for b.Loop() {
 				username := f.user
 				if test.createUser {
-					username = utils.GenerateLocalUsername(b)
+					username = testutils.GenerateLocalUsername(b)
 					b.Cleanup(func() { _, _ = host.UserDel(username) })
 				}
